chore: 반려동물 몸무게 자료형 변경에 따른 mapper 수정

+++ src/main/java/com/kau/capstone/domain/pet/util/PetMapper.java
@@ -31,7 +31,7 @@ public class PetMapper {
             .petType(pet.getPetType())
             .gender(pet.getGender())
             .name(pet.getName())
-            .weight(pet.getWeight())
+            .weight((float) pet.getWeight())
             .imageUrl(pet.getImageUrl())
             .build();
     }
